Integrate Message Broker
Use RabbitMQ as a Message Broker for the purpose of making our system
loosely coupled.

Two ways of communication were adopted :

1. Sending messages to a Queue
To quickly check and confirm that our system is working properly ,
sending message started by routing directly to a queue.

This is handled by message-channel
2. Sending Message to an Exchange

For proper messaging , we should send messages to an exchange using
a routing key that will determin how  a queue is binded